feat(render): add parsed messages and decode code
Rename RustWireMessageGenerator.xtend -> RustMessageGenerator.xtend.

Add code generation for parsed_message.rs that has a struct for each message with fields, including String for the strings sent in the message. Add a decode() method to instantiate each message struct from a buffer containing a message of that type.

Also fix warning in Cargo Generator due to having profile.release in
internal crates. We should have removed it when we started compiling from
the root workspace, since that config belongs there.